Remove recycle bin folder support
Currently, UGENE does not use folders at all. I want to keep the folders support and improve it in the next patches, but the recycle bin is implemented not correctly - so I removing it.

The correct implementation will use a 100% separate location for recycled objects, so there will be no need at all to check if the object/folder is in the recycle bin across the code.